---======First Commit======--- All Forms and Classes are created and ready to be used as intended.
====View Quotes Form====
All buttons and menu strip items are implemented and in working condition, this includes the instantiation of the appropriate classes.

====Add Pump Form====
Form Activation implemented to allow the Edit or the view of a specific pump that was passed to the Form from the originating Form.
Form Activation also determines when a new pump should be created by using the pass class variables.

==Menu Stip==
No functionality was implemented besides the closure of the form by pressing the "Close" menu strip item.

==Add/Edit Button==
Implemented Add New Pump functionality. Editing functionality implementation is still outstanding.

++++++ Compiles and Runs - Limited functionality++++++