---======Third Commit======---
All Forms and Classes are created and ready to be used as intended.

All Forms' Anchor properties updated to allow the user to adjust a Form's size according to their needs

====View Quotes Form====
All buttons and menu strip items are implemented and in working condition, this includes the instantiation of the appropriate classes.

====Add Pump Form====
Form Activation implemented to allow the Edit or the view of a specific pump that was passed to the Form from the originating Form.
Form Activation also determines when a new pump should be created by using the pass class variables.

==Menu Stip==
No functionality was implemented besides the closure of the form by pressing the "Close" menu strip item.

==Add/Edit Button==
Implemented Add New Pump functionality as well as Editing a passed pump object's information. Extra functionality added to ensure null objects are not accepted and prevent possible errors related to this.

====Add Part Form====
Form Activation implemented to allow the Edit or the view of a specific pump that was passed to the Form from the originating Form.
Form Activation also determines when a new Part should be created by using the pass class variables.

==Menu Stip==
New "Import Batch" Menu strip item added to the "File" Menu strip item, this allows the user to import a CSV file to add Parts to the system rather than adding them individually.

New "Reset Input"  menu strip item added to the "File" Menu strip item, this allows the user to reset the screen and enter new information.

==Add/Edit Part Button==
"Add Pump" button implemented and works as expected.

====View Parts Form====
Form Activation implemented to list the Mandatory Pars first the Non-Mandatory parts

==Menu Stip==
No Functionality implemented besides the Close menu strip item.

==Add Part Button==
This button opens the "Add/Edit Part" form and hides the current form.

==Remove Part Button==
Removes a selected part from the list. This is not workin correctly as of yet, I failed to realise that the data grid view can sort its items on it own which in turn causes the wrong items to be removed in certain cases.

==Update Part Button==
No Functionality as of yet.

++++++ Compiles and Runs - Limited functionality ++++++